requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
debtcollector==1.22.0
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.3
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
keystoneauth1==3.18.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
netaddr==0.7.19
netifaces==0.10.9
os-service-types==1.7.0
oslo.i18n==3.25.0
oslo.utils==3.42.0
pbr==5.4.4
prettytable==0.7.2
pyparsing==2.4.5
python-cinderclient==5.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
simplejson==3.17.0
six==1.13.0
SQLAlchemy==1.3.11
stevedore==1.31.0
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
